Exempt internal/generated links from htmlproofer
The links on the post overview page for posts we host are automatically
generated so they will never fail, we can ignore these. If we don't add
creation time the external resource won't exist yet and Travis will
register a build failure.

Similarly for links we've defined in our site navigation, if we're
addind new pages those pages aren't there yet so the tests would fail.