Use javascript to dynamically get stats for the main page (#60)
Before this, the main page had hand-curated data that was rarely
updated.

Getting repo count and member count is pretty easy. I am leaving pull
request count for later work because it is not easy.

Note that when I look at 'members' on github I see 60+ members, but that
is because I am an admin. When not authenticated I only see the 29
people that have publicized their membership. This is true of the
website as well, so it will sortof look like our membership dropped.

Changed contributors to collaborators in line with the not-quite-yet
merged governance PR.