dont load js to contact github api
Normally, we render the stats on our website live from the github API.
This triggers multiple API calls from the visitors workstation to
github. Also github does not display members with a private membership
in that API call. So the list of displayed contributors is way smaller
than it actually is.