[WIP] the start of a set of reference architectures
Do not merge this as is. It's not ready to go live on the site.
You'll want to use the `Checkout via command line` instructions on the
PR to check this out locally so you can tweak the graph and see it live.

Check out the PR, run the site locally with `jekyll serve`, tweak and
improve as needed.

It's a reasonable example, for sure, but the purpose of this PR is
simply to create the framework and demonstrate an example so that
you can more easily put together the reference yourself.