Attempt to improve the website
Yeah, I am not a web developer!  We have inconsistencies and blatant UI
issues that this PR attemps to solve.

We depend on outdated third party "things" (e.g. Material Design for
Bootstrap) that should probably be updated so that we can follow the
current documentation, but that would be more effort that what I can
offer at the moment.