Code of conduct: Update version 1.4 -> 2.1
We use the code of conduct from https://www.contributor-covenant.org/.
This PR updates our old version of 1.4 to 2.1.

I tried to keep the diff as readable as possible and ignored markdown
linting to make reviewing it easier. This is a direct copy of
https://www.contributor-covenant.org/version/2/1/code_of_conduct/
with one exception. We previously had:

```
Our maintainers are elected on a yearly basis as the project management
committee. You can find an up to date list of those people [on our site][pmc]
```

in the `Enforcement Responsibilities` section. I kept it.