revamps navigation menus by adding dropdowns
This takes the very busy navigation bars and reorganizes them into
dropdown menus. This gives us a lot more open space, but also lets have
a more intuitive navigation structure.

- moves the "connect with us" links from the footer into a nav menu.
  - also creates a `/connect` page that better explains what's bridged
    and how you can choose protocols to connect with.
- adds a Project Overview page with links to github repo lists
- configures the entire nav structure in `_config.yml` instead of some
  being hardcoded in the html
- adds the ability to add icons to any menu item.
  - if the icons aren't appreciated, just delete from the config
  - updated to current FontAwesome so the online search is usable

I did -not- make the money icon spin around, but I was tempted to.

If you like the overall look & feel, but want to adjust the menu
organization, can I suggest that we merge this and then contribute a
separate PR to reorganize?

This does not yet add any of the OpenPuppet pages. I'll do that in a
separate PR and then add navigation after we decide on a name.