Fix and improve monthly sync information
Time Zones are a PITA, the meeting incidently happen at 15:30 UTC at the
moment, but it is in reality scheduled at 16:30 CET (UTC+1) during the winter
and 16:30 CEST (UTC+2) when daylight sayving take place during the
summer.  Fix the indication so that it matches reality.

Because converting from one time-zone to an other is something nobody
like to do, add some JavaScript code to do it for us and display when
the next monthly sync will take place.  For this we rely on moment.js to
do the math (as everybody knows, JS sucks and it lacks some time-zone
functions) and format the time of the event conveniently for the user in
their time zone (with formatting such as "Tomorrow at 5:30 AM" when the
date is near).

Last, but not least, provide a .ics file with the same information one
more time so that people can easily choose what they prefer.  Compared
to the event in the Puppetlabs calendar, this one supposely has the
correct time-zone and should avoid the mess that happened during the
last daylight saving change where US and EU where not changing the same
week.