[draft] Adding community fork section
***Please do not merge as-is. We need the approved name first.***

This is a first draft of what the "OpenPuppet" (name TBD) section could
look like. It adds a small subsection tab-bar and refactors the sponsor
cards into a reusable component. It's not at all ready to merge, I just
wanted to get it to a point that others could poke at it and offer
feedback or improvements. Many of the images are just placeholders until
we get a name and some branding.

To preview it, check out this pull request, then

```
$ bundle install
$ bundle exec jekyll serve
```