Refactor how pages describe interaction
Connect had Slack/IRC/Mailing Lists
Contributing had short versions of the Slack/IRC/Mailing List, but also
had the Monthly Sync.  This moves each of those into its own include and
then puts all of them on both pages.

Replace include type

include_relative blocks ../ and include
automatically looks at the _includes directory
which is where all these files look to exist.

Fix the link to the ics

Build the pages and attach to PR as an artifact

This should provide people a preview that they can pull locally and look
at.  The artifacts are set to last 5 days, as to not burn storage.

Remove the duplicate mention of the channel

Signed-off-by: Jason Straw <jason.straw@gmail.com>