Add a checklist (summarizing the full list)
This makes a clean markdown list to copy/paste with the checkboxes
that github issues can handle.  It rewrites some of the entries to
be shorter to make it less complex to read, while linking back to
the website for reference material.

Additionally, Remove the forge approved module step.  As noted in
IRC, this should be a PMC decision, reviews will be requested from
PMC members.