Adds a basic VPAT document
See https://en.wikipedia.org/wiki/Voluntary_Product_Accessibility_Template
for information about the VPAT standard.

This describes the accessibility features of OpenVox. It does -NOT- add
it to the navigation menu at this time, so the URL will need to be
shared directly with users who require it. We can figure out later how
we want to display or link to this.

The link will be http://voxpupuli.org/openvox/vpat/