Update software support page with tooling recommendations
This reworks the page slightly to broaden its scope. Now it covers the
guidelines for choosing languages and frameworks for projects as well as
just OS/platform support timelines.

Note: this does not yet replace Puppet with OpenVox. I'd love it if
someone took the time to write that up.

Also, please do continue expanding the guidelines. This is not intended
to be the end point for the page!